moved seed into seed.rb, added hooks deploystep

+++ db/migrate/20140203075119_rem_sudo_auth_from_app.rb
@@ -1,9 +1,10 @@
 class RemSudoAuthFromApp < ActiveRecord::Migration
   def up
     dsto = DeployStepTypeOption.find_by name: "has_sudo"
-    DeployStep.destroy_all(deploy_step_type_option_id: dsto.id)
-    dsto.destroy
+    if dsto
+      DeployStep.destroy_all(deploy_step_type_option_id: dsto.id)
+      dsto.destroy
+    end
     DeployStepType.destroy_all(name: "super_powers")
   end
-
 end
